0.1.0
Added /config/connection.js to require Sequelize package and create a new instance of Sequelize then export, Added /db/schema.sql to drop db name if exists and create db named devblog_db, Added /models/user.js created a user model, that has a method verify pass to compare the bcrypt encryptoion for auth set up user model to have not allow null values in username or password or email and added a hook so that beforeCreate -ing the user the user.password will use bcrypt to has the newuser objects password property. then exported to index.js. added /seeds/userData.js to seed a sample user with User.bulkCreate and seed array of objects as an arg. then exported to index.js where it can be all seeds and be called, added server.js that requires needed packages, created a cookie, used express middleware with the session that uses the cookie as an arg then wrapped the app.listen in a sequelize.sync(). updated readme.md to show that repo is currently under construction.