0.4.0
Large update: Updated controller to have a api folder and changed data related enpoints to use that route. Added a login route that validates with the models method verifyPass(), added a logout route that logs out the user with req.session destroy, added newPost route for new posts to be added to the blog, added new feature where the username will equal the req.session.username which was implemented by setting it straight from when a user is created but that only works in that particular session instance so another method was added for log in to call a new method on the module which returns the username of the use that logged in and save it in session data so that it can be used to tag posts as expected. updated models/user.js to have that method. updated controller/index.js to use the /api endpoint for routes in the api folder. updated all client side fetches so that if they recieved response status of 200 or 204 they will return to their relative page in most instances the home page this was done to fix a bug cause by using res.send and res.redirect simultaniously. public/js/logout was added to make a post fetch to logout to destroy the session,  public/js/newpost was created to make a client side fetch to the post method which creates a new Post and then updates it to the homepage. homepage was fitted with 2 script tags, public/logout.js and public/newpost.js to handle those routes, Main js added conditionals to decided if script tags are necissary, or if the option to be able to make a post above the blogs section will be show. it only has that capacity for logged in users.