0.5.0
added api/commentroute.js that listens for a post request then calls create on the comment model, with an object as an argument that uses the data sent from the client side post fetch with req.body.propname. if it is sucessfuly server will respond with 200 and data, if fail, 500 + errcode.  Added get route to controller/homepage-routes.js, that returns the data needed for 1 post object that endpoint will be selected when the post is clicked the client uses and href=/post/(id) this enpoint uses Post.findbypk and uses req.params.id + includes clause as arguments then, if all is sucessful response is 200+ render post handlebars view + data object + loggedin status. added the commentRoutes to use the api endpoint in index.js. in models created a comment.js model that has id, comment_body, post_id and username(which will be assigned with req.session.username, which is assigned to the session on login or signup. index.js was updated to create the relationship between comment and posts as comment belogs to post and posts has many comments. updated style.css to effect the new comments with styles. added public/js/comment.js to make the client side fetch that is needed for the controller/api/commentroutes.js this uses a eventlistner with the wind.location.href and spliting /manipulating the data to get the post_id that can be used when the new comment is created. then if all is well and status is 200 user is redirected to the endpoint by using the same post_id. created seeds/comment.js to seed sample data for testing on build however now since the username is set by session, they are obsolete. updated index.js to seedComments with seedall function. updated views/layouts/main.handlebars removed the login state that rendered a form to make a new post since it was showing up on all pages when logged in. moved this to the homepage.handlebars and added a script tag for the public/js/comment.js client side fetch. added a views/post.handlebars for view off the single clicked post, if the user is logged in they will be givin the option to make a comment, which will trigger the comment.js clientside post fetch.