0.6.0
Updated controllers/api/commentroutes.js to fill a date feild now thats saved on creation of the comment as the time of creation. updated controllers/homepage-routes.js to have dashboard routes wich will eventually be isolated in their own space, but for now get route was created for /dashboards, uses conditional to check if user is logged in, If not, user is redirected to login otherwise, All posts By the session user are shown with the options to update or delete a route. using the newly added update route that await Post.update with the arg of the req body from the client fetch and where clause for where to update that is where id = req.body.post_id.  if all went well status code 200 + data else, status 500 + errcode. and the newly added delete route which is simple in nature and just uses destroy on model with a where clause set to where id = req.body.post_id. Models/comment.js has been changed to accept a date field, Models/post.js has been changed to accept a date field, models/index.js has been updated to fix a bug where a post would not be deleted if a comment was on it, Added ondelete CASCADE to fix. added a /public/delpost.js to make the client side fetch by added and event listener to each of the elements, and then using the event target id, as the body for the route to use in its where clause. Created an Public/js/updatepost.js which uses the same queryselectorAll to return an array of objects using that class and then creating an event listener for each on with a for of loop, then this calls a function with the argument of the e.target.id  this function uses a doc.queryselector on the class updateform, then sets the update forms css text to display block since by nature it is a hidden element. then adds and event listener to the new button of the form elem this is where all data and variables from the form are set and agument from the event listener is passed in and used then a fetch put is made for the route to use the body to update the data. if this is cuessfull and res.status is 200 client uses window.location.href to redirect users back to their dashboard. Views updates: /layouts/main.handlebars was updated with scripttags for updatepost.js and delpost.js.  views/dashboard.handlebars was added to that users could see only posts by themselves and update/delete these posts. homepage.handlebars, and post.handlebars and dashboard.handlebars were all updated to use new date_format from utils, and the newly added date field for posts and comments.