1.2.2
fixed bug in the way update form displays by changing queryselector to query selector all since it always returned the first item the form would always open after the first element no matter which post was selected for update then refrenced the parentNode which is the form and set css style to display block.  Fixed bug in views where logout button was not working in the post views added ../ to script tag since the directory is /post/id.